Fix: Fix build errors
